+++ src/JIM.Worker/Processors/SyncTaskProcessorBase.cs
@@ -2212,6 +2212,8 @@ public abstract class SyncTaskProcessorBase
             attributeChange = new MetaverseObjectChangeAttribute
             {
                 Attribute = metaverseObjectAttributeValue.Attribute,
+                AttributeName = metaverseObjectAttributeValue.Attribute.Name,
+                AttributeType = metaverseObjectAttributeValue.Attribute.Type,
                 MetaverseObjectChange = metaverseObjectChange
             };
             metaverseObjectChange.AttributeChanges.Add(attributeChange);